+++ backend/main.py
@@ -2,17 +2,14 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 
-#FastAPI instance
-#================================#
+# FastAPI instance
 app = FastAPI(
-    version="0.0.1",
+    version="0.0.3",
     title="Worganizer",
     description="The Ultimate Vocabulary Organizer For Language Learning"
 )
-#================================#
 
-#CORS treatment
-#================================#
+# CORS treatment
 origins = [
     "*"
 ]
@@ -23,14 +20,11 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-#================================#
 
-#Routes
-#================================#
-from src.routes import users
+# Routes
+from src.routes import users, words
 
-#Root route of the backend
+# Root route of the backend
 @app.get('/', tags=['Root Route'])
 def worganizer_root_route():
     return {'Worganizer': 'The Ultimate Vocabulary Organizer For Language Learning'}
-#================================#